move tree view from the expression builder to a dedicated class
