+++ src/gui/CMakeLists.txt
@@ -383,6 +383,7 @@ SET(QGIS_GUI_SRCS
   qgsexpressionlineedit.cpp
   qgsexpressionselectiondialog.cpp
   qgsexpressionstoredialog.cpp
+  qgsexpressiontreeview.cpp
   qgsextentgroupbox.cpp
   qgsextentwidget.cpp
   qgsexternalresourcewidget.cpp
@@ -602,6 +603,7 @@ SET(QGIS_GUI_HDRS
   qgsexpressionhighlighter.h
   qgsexpressionlineedit.h
   qgsexpressionselectiondialog.h
+  qgsexpressiontreeview.h
   qgsextentgroupbox.h
   qgsextentwidget.h
   qgsexternalresourcewidget.h
