+++ src/gui/qgsexpressionbuilderwidget.h
@@ -17,114 +17,22 @@
 #define QGSEXPRESSIONBUILDER_H
 
 #include <QWidget>
-#include "qgis_sip.h"
+#include <QStandardItemModel>
+#include <QSortFilterProxyModel>
+
 #include "ui_qgsexpressionbuilder.h"
+
+#include "qgis_sip.h"
+#include "qgis_gui.h"
 #include "qgsdistancearea.h"
 #include "qgsexpressioncontext.h"
 #include "qgsexpression.h"
+#include "qgsexpressiontreeview.h"
 
-#include "QStandardItemModel"
-#include "QStandardItem"
-#include "QSortFilterProxyModel"
-#include "QStringListModel"
-#include "qgis_gui.h"
 
 class QgsFields;
 class QgsExpressionHighlighter;
 class QgsRelation;
-
-/**
- * \ingroup gui
- * An expression item that can be used in the QgsExpressionBuilderWidget tree.
-  */
-class GUI_EXPORT QgsExpressionItem : public QStandardItem
-{
-  public:
-    enum ItemType
-    {
-      Header,
-      Field,
-      ExpressionNode
-    };
-
-    QgsExpressionItem( const QString &label,
-                       const QString &expressionText,
-                       const QString &helpText,
-                       QgsExpressionItem::ItemType itemType = ExpressionNode )
-      : QStandardItem( label )
-    {
-      mExpressionText = expressionText;
-      mHelpText = helpText;
-      mType = itemType;
-      setData( itemType, ITEM_TYPE_ROLE );
-    }
-
-    QgsExpressionItem( const QString &label,
-                       const QString &expressionText,
-                       QgsExpressionItem::ItemType itemType = ExpressionNode )
-      : QStandardItem( label )
-    {
-      mExpressionText = expressionText;
-      mType = itemType;
-      setData( itemType, ITEM_TYPE_ROLE );
-    }
-
-    QString getExpressionText() const { return mExpressionText; }
-
-    /**
-     * Gets the help text that is associated with this expression item.
-      *
-      * \returns The help text.
-      */
-    QString getHelpText() const { return mHelpText; }
-
-    /**
-     * Set the help text for the current item
-      *
-      * \note The help text can be set as a html string.
-      */
-    void setHelpText( const QString &helpText ) { mHelpText = helpText; }
-
-    /**
-     * Gets the type of expression item, e.g., header, field, ExpressionNode.
-      *
-      * \returns The QgsExpressionItem::ItemType
-      */
-    QgsExpressionItem::ItemType getItemType() const { return mType; }
-
-    //! Custom sort order role
-    static const int CUSTOM_SORT_ROLE = Qt::UserRole + 1;
-    //! Item type role
-    static const int ITEM_TYPE_ROLE = Qt::UserRole + 2;
-    //! Search tags role
-    static const int SEARCH_TAGS_ROLE = Qt::UserRole + 3;
-
-  private:
-    QString mExpressionText;
-    QString mHelpText;
-    QgsExpressionItem::ItemType mType;
-
-};
-
-/**
- * \ingroup gui
- * Search proxy used to filter the QgsExpressionBuilderWidget tree.
-  * The default search for a tree model only searches top level this will handle one
-  * level down
-  */
-class GUI_EXPORT QgsExpressionItemSearchProxy : public QSortFilterProxyModel
-{
-    Q_OBJECT
-
-  public:
-    QgsExpressionItemSearchProxy();
-
-    bool filterAcceptsRow( int source_row, const QModelIndex &source_parent ) const override;
-
-  protected:
-
-    bool lessThan( const QModelIndex &left, const QModelIndex &right ) const override;
-};
 
 
 /**
@@ -145,22 +53,23 @@ class GUI_EXPORT QgsExpressionBuilderWidget : public QWidget, private Ui::QgsExp
 
     /**
      * Sets layer in order to get the fields and values
-      * \note this needs to be called before calling loadFieldNames().
-      */
+     * \note this needs to be called before calling loadFieldNames().
+     */
     void setLayer( QgsVectorLayer *layer );
 
     /**
-     * Loads all the field names from the layer.
-      * @remarks Should this really be public couldn't we just do this for the user?
-      */
-    void loadFieldNames();
+     * Returns the current layer or a nullptr.
+     */
+    QgsVectorLayer *layer() const;
 
-    void loadFieldNames( const QgsFields &fields );
+    //! \deprecated since QGIS 3.14 this is now done automatically
+    Q_DECL_DEPRECATED void loadFieldNames( const QgsFields &fields = QgsFields() ) {Q_UNUSED( fields )}
 
     /**
      * Loads field names and values from the specified map.
      *  \note The field values must be quoted appropriately if they are strings.
      *  \since QGIS 2.12
+     * \deprecated use setLayer() and expressionTree()->
      */
     void loadFieldsAndValues( const QMap<QString, QStringList> &fieldValues );
 
@@ -200,63 +109,51 @@ class GUI_EXPORT QgsExpressionBuilderWidget : public QWidget, private Ui::QgsExp
 
     /**
      * Sets the expression context for the widget. The context is used for the expression
-     * preview result and for populating the list of available functions and variables.
+     * preview result and to populate the list of available functions and variables.
      * \param context expression context
      * \see expressionContext
      * \since QGIS 2.12
      */
     void setExpressionContext( const QgsExpressionContext &context );
 
-    /**
-     * Registers a node item for the expression builder.
-      * \param group The group the item will be show in the tree view.  If the group doesn't exist it will be created.
-      * \param label The label that is show to the user for the item in the tree.
-      * \param expressionText The text that is inserted into the expression area when the user double clicks on the item.
-      * \param helpText The help text that the user will see when item is selected.
-      * \param type The type of the expression item.
-      * \param highlightedItem set to TRUE to make the item highlighted, which inserts a bold copy of the item at the top level
-      * \param sortOrder sort ranking for item
-      * \param icon custom icon to show for item
-      * \param tags tags to find function
-      */
-    void registerItem( const QString &group, const QString &label, const QString &expressionText,
-                       const QString &helpText = QString(),
-                       QgsExpressionItem::ItemType type = QgsExpressionItem::ExpressionNode,
-                       bool highlightedItem = false, int sortOrder = 1,
-                       QIcon icon = QIcon(),
-                       const QStringList &tags = QStringList() );
-
     bool isExpressionValid();
 
     /**
      * Adds the current expression to the given \a collection.
      * By default it is saved to the collection "generic".
+     * \deprecated since QGIS 3.14 use expressionTree()->saveRecent() instead
      */
-    void saveToRecent( const QString &collection = "generic" );
+    Q_DECL_DEPRECATED void saveToRecent( const QString &collection = "generic" );
 
     /**
      * Loads the recent expressions from the given \a collection.
      * By default it is loaded from the collection "generic".
+     * \deprecated since QGIS 3.14 use expressionTree()->loadRecent() instead
      */
-    void loadRecent( const QString &collection = QStringLiteral( "generic" ) );
+    Q_DECL_DEPRECATED void loadRecent( const QString &collection = QStringLiteral( "generic" ) );
+
+    QgsExpressionTreeView *expressionTree() const;
 
     /**
      * Loads the user expressions.
+     * \deprecated since QGIS 3.14 use expressionTree()->loadUserExpressions() instead
      * \since QGIS 3.12
      */
-    void loadUserExpressions( );
+    Q_DECL_DEPRECATED void loadUserExpressions();
 
     /**
      * Stores the user \a expression with given \a label and \a helpText.
+     * \deprecated since QGIS 3.14 use expressionTree()->saveToUserExpressions() instead
      * \since QGIS 3.12
      */
-    void saveToUserExpressions( const QString &label, const QString expression, const QString &helpText );
+    Q_DECL_DEPRECATED void saveToUserExpressions( const QString &label, const QString expression, const QString &helpText );
 
     /**
      * Removes the expression \a label from the user stored expressions.
+     * \deprecated since QGIS 3.14 use expressionTree()->removeFromUserExpressions() instead
      * \since QGIS 3.12
      */
-    void removeFromUserExpressions( const QString &label );
+    Q_DECL_DEPRECATED void removeFromUserExpressions( const QString &label );
 
     /**
      * Creates a new file in the function editor
@@ -287,8 +184,9 @@ class GUI_EXPORT QgsExpressionBuilderWidget : public QWidget, private Ui::QgsExp
      * Returns a pointer to the dialog's function item model.
      * This method is exposed for testing purposes only - it should not be used to modify the model.
      * \since QGIS 3.0
+     * \deprecated since QGIS 3.14
      */
-    QStandardItemModel *model();
+    Q_DECL_DEPRECATED QStandardItemModel *model();
 
     /**
      * Returns the project currently associated with the widget.
@@ -388,31 +286,17 @@ class GUI_EXPORT QgsExpressionBuilderWidget : public QWidget, private Ui::QgsExp
     void editSelectedUserExpression();
 
     /**
-     * Create the expressions JSON document storing all the user expressions to be exported.
-     * \returns the created expressions JSON file
-     * \since QGIS 3.14
-     */
-    QJsonDocument exportUserExpressions();
-
-    /**
-     * Load and permanently store the expressions from the expressions JSON document.
-     * \param expressionsDocument the parsed expressions JSON file
-     * \since QGIS 3.14
-     */
-    void loadExpressionsFromJson( const QJsonDocument &expressionsDocument );
-
-    /**
      * Returns the list of expression items matching a \a label.
      * \since QGIS 3.12
+     * \deprecated since QGIS 3.14 use expressionTree()->findExpressions instead
      */
     const QList<QgsExpressionItem *> findExpressions( const QString &label );
 
 
   private slots:
     void indicatorClicked( int line, int index, Qt::KeyboardModifiers state );
-    void showContextMenu( QPoint );
     void setExpressionState( bool state );
-    void currentChanged( const QModelIndex &index, const QModelIndex & );
+    void expressionTreeItemChanged( QgsExpressionItem *item );
     void operatorButtonClicked();
     void btnRun_pressed();
     void btnNewFile_pressed();
@@ -431,9 +315,8 @@ class GUI_EXPORT QgsExpressionBuilderWidget : public QWidget, private Ui::QgsExp
      */
     void importUserExpressions_pressed();
     void cmbFileNames_currentItemChanged( QListWidgetItem *item, QListWidgetItem *lastitem );
-    void expressionTree_doubleClicked( const QModelIndex &index );
+    void insertExpressionText( const QString &text );
     void txtExpressionString_textChanged();
-    void txtSearchEdit_textChanged();
     void txtSearchEditValues_textChanged();
     void lblPreview_linkActivated( const QString &link );
     void mValuesListView_doubleClicked( const QModelIndex &index );
@@ -469,67 +352,30 @@ class GUI_EXPORT QgsExpressionBuilderWidget : public QWidget, private Ui::QgsExp
     void showEvent( QShowEvent *e ) override;
 
   private:
+    class ExpressionTreeMenuProvider : public QgsExpressionTreeView::MenuProvider
+    {
+      public:
+        ExpressionTreeMenuProvider( QgsExpressionBuilderWidget *expressionBuilderWidget )
+          : QgsExpressionTreeView::MenuProvider()
+          , mExpressionBuilderWidget( expressionBuilderWidget ) {}
+
+        QMenu *createContextMenu( QgsExpressionItem *item ) override;
+
+      private:
+        QgsExpressionBuilderWidget *mExpressionBuilderWidget;
+    };
+
     int FUNCTION_MARKER_ID = 25;
     void createErrorMarkers( QList<QgsExpression::ParserError> errors );
     void createMarkers( const QgsExpressionNode *node );
     void clearFunctionMarkers();
     void clearErrors();
     void runPythonCode( const QString &code );
-    void updateFunctionTree();
     void fillFieldValues( const QString &fieldName, int countLimit, bool forceUsedValues = false );
     bool formatterCanProvideAvailableValues( const QString &fieldName );
     QString getFunctionHelp( QgsExpressionFunction *function );
     QString loadFunctionHelp( QgsExpressionItem *functionName );
     QString helpStylesheet() const;
-
-    void loadExpressionContext();
-
-    //! Loads current project relations names/id into the expression help tree
-    void loadRelations();
-
-    //! Loads current project layer names/ids into the expression help tree
-    void loadLayers();
-
-    /**
-     * Registers a node item for the expression builder, adding multiple items when the function exists in multiple groups
-      * \param groups The groups the item will be show in the tree view.  If a group doesn't exist it will be created.
-      * \param label The label that is show to the user for the item in the tree.
-      * \param expressionText The text that is inserted into the expression area when the user double clicks on the item.
-      * \param helpText The help text that the user will see when item is selected.
-      * \param type The type of the expression item.
-      * \param highlightedItem set to TRUE to make the item highlighted, which inserts a bold copy of the item at the top level
-      * \param sortOrder sort ranking for item
-      * \param tags tags to find function
-      */
-    void registerItemForAllGroups( const QStringList &groups, const QString &label, const QString &expressionText,
-                                   const QString &helpText = QString(),
-                                   QgsExpressionItem::ItemType type = QgsExpressionItem::ExpressionNode,
-                                   bool highlightedItem = false, int sortOrder = 1, const QStringList &tags = QStringList() );
-
-    /**
-     * Returns a HTML formatted string for use as a \a relation item help.
-     */
-    QString formatRelationHelp( const QgsRelation &relation ) const;
-
-    /**
-     * Returns a HTML formatted string for use as a \a layer item help.
-     */
-    QString formatLayerHelp( const QgsMapLayer *layer ) const;
-
-    /**
-     * Returns a HTML formatted string for use as a \a recent \a expression item help.
-     */
-    QString formatRecentExpressionHelp( const QString &label, const QString &expression ) const;
-
-    /**
-     * Returns a HTML formatted string for use as a \a user \a expression item help.
-     */
-    QString formatUserExpressionHelp( const QString &label, const QString &expression, const QString &description ) const;
-
-    /**
-     * Returns a HTML formatted string for use as a \a variable item help.
-     */
-    QString formatVariableHelp( const QString &variable, const QString &description, bool showValue, const QVariant &value ) const;
 
     /**
      * Will be set to TRUE if the current expression text reported an eval error
@@ -549,37 +395,20 @@ class GUI_EXPORT QgsExpressionBuilderWidget : public QWidget, private Ui::QgsExp
 
     void loadFieldValues( const QVariantMap &values );
 
-    void loadFieldsAndValues( const QMap<QString, QVariantMap> &fieldValues );
-
-    /**
-     * Display a message box to ask the user what to do when an expression
-     * with the same \a label already exists. Answering "Yes" will replace
-     * the old expression with the one from the file, while "No" will keep
-     * the old expression.
-     * \param isApplyToAll whether the decision of the user should be applied to any future label collision
-     * \param isOkToOverwrite whether to overwrite the old expression with the new one in case of label collision
-     * \param label the label of the expression
-     * \param oldExpression the old expression for a given label
-     * \param newExpression the new expression for a given label
-     * \since QGIS 3.14
-     */
-    void showMessageBoxConfirmExpressionOverwrite( bool &isApplyToAll, bool &isOkToOverwrite, const QString &label, const QString &oldExpression, const QString &newExpression );
-
-    bool mAutoSave = true;
-    QString mFunctionsPath;
-    QgsVectorLayer *mLayer = nullptr;
-    std::unique_ptr<QStandardItemModel> mModel;
     // Will hold items with
     // * a display string that matches the represented field values
     // * custom data in Qt::UserRole + 1 that contains a ready to use expression literal ('quoted string' or NULL or a plain number )
     std::unique_ptr<QStandardItemModel> mValuesModel;
     std::unique_ptr<QSortFilterProxyModel> mProxyValues;
-    std::unique_ptr<QgsExpressionItemSearchProxy> mProxyModel;
-    QMap<QString, QgsExpressionItem *> mExpressionGroups;
+
+    ExpressionTreeMenuProvider *mExpressionTreeMenuProvider = nullptr;
+
+    bool mAutoSave = true;
+    QString mFunctionsPath;
+    QgsVectorLayer *mLayer = nullptr;
     QgsExpressionHighlighter *highlighter = nullptr;
     bool mExpressionValid = false;
     QgsDistanceArea mDa;
-    QString mRecentKey;
     QMap<QString, QVariantMap>  mFieldValues;
     QgsExpressionContext mExpressionContext;
     QPointer< QgsProject > mProject;
@@ -587,7 +416,6 @@ class GUI_EXPORT QgsExpressionBuilderWidget : public QWidget, private Ui::QgsExp
     bool mParserError = true;
     // Translated name of the user expressions group
     QString mUserExpressionsGroupName;
-    QStringList mUserExpressionLabels;
 };
 
 // clazy:excludeall=qstring-allocations
